requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
Django==1.11.7
djangorestframework==3.7.7
factory-boy==2.11.1
Faker==0.9.2
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.8.0
py==1.7.0
pytest==3.9.2
pytest-django==3.4.3
python-dateutil==2.7.3
pytz==2017.3
six==1.11.0
text-unidecode==1.2
